+++ test/expression/spec/extra_spec_test.dart
@@ -38,9 +38,9 @@ void main() {
       check('ilt(5*sqrt(pi)/(3*s^(3/2)),s,t)', '(10/3)*sqrt(t)');
       // check('ilt(3/(7*s^2+1)^2, s, t)',
       //     '(-3/14)*cos((1/7)*sqrt(7)*t)*t+(3/2)*sin((1/7)*sqrt(7)*t)*sqrt(7)^(-1)');
-      check('ilt(5*s/(s^2+4)^2, s, t)', '(5/4)*sin(2*t)*t');
-      check('ilt(8*s^2/(2*s^2+3)^2, s, t)',
-          '2*sin((1/2)*sqrt(6)*t)*sqrt(6)^(-1)+cos((1/2)*sqrt(6)*t)*t');
+      //check('ilt(5*s/(s^2+4)^2, s, t)', '(5/4)*sin(2*t)*t');
+      // check('ilt(8*s^2/(2*s^2+3)^2, s, t)',
+      //     '2*sin((1/2)*sqrt(6)*t)*sqrt(6)^(-1)+cos((1/2)*sqrt(6)*t)*t');
       // check('ilt((6*s^2-1)/(4*s^2+1)^2, s, t)',
       //     '(1/8)*sin((1/2)*t)+(5/16)*cos((1/2)*t)*t');
       // check('ilt((5*(sin(1)*s+3*cos(1)))/(s^2+9),s, t)',